// src/app/@features/admin/dashboard/models/adminDashboardInsights.ts
export class AdminDashboardInsights {
  totalAdvertisers: number;
  totalPromoters: number;
  pendingCommissionsCount: number;
  pendingCommissionsAmount: number | string;
  approvedCommissionsCount: number;
  approvedCommissionsAmount: number | string;
  rejectedCommissionsCount: number;
  rejectedCommissionsAmount: number | string;
  paidCommissionsAmount: number | string;
  totalBalance: number | string;
  totalProfit: number | string;
}